- updated event bass worker logic -  moved navigation guards to a separate directory - custom directives moved to a separate directory
