+++ src/components/filters/components/RangeFilter.vue
@@ -64,16 +64,11 @@ export default {
 </script>
 
 <style scoped lang="scss">
-$color-148783: #148783;
-$color-1ab2ad: #1ab2ad;
-$color-e9fcfb: #e9fcfb;
-$color-f9f9fc: #f9f9fc;
-$color-dfdeed: #dfdeed;
 .filter-style {
-  --slider-connect-bg: $color-148783;
+  --slider-connect-bg: #148783;
   --slider-height: 2px;
-  --slider-handle-border: 4px solid $color-1ab2ad;
-  --slider-handle-shadow: 0.5px 0.5px 2px 1px $color-e9fcfb;
+  --slider-handle-border: 4px solid #1ab2ad;
+  --slider-handle-shadow: 0.5px 0.5px 2px 1px #e9fcfb;
 }
 .b-range-filter__slider-wrapper {
   display: flex;
@@ -87,29 +82,29 @@ $color-dfdeed: #dfdeed;
     font-weight: 400;
     font-size: 12px;
     line-height: 16px;
-    color: $--b-main-black-color;
+    color: #262541;
     padding: 0px 4px;
-    background: $color-f9f9fc;
+    background: #f9f9fc;
     border-radius: 4px;
     z-index: 1;
   }
   .b-range-filter__left-side {
-    background: $--b-main-white-color;
+    background: #ffffff;
     padding: 0 12px;
     width: 100%;
     display: flex;
     align-items: center;
-    border: 1px solid $color-dfdeed;
+    border: 1px solid #dfdeed;
     border-radius: 6px 0 0 6px;
     .b-range-filter__slider {
       width: 100%;
     }
   }
   .b-range-filter__right-side {
-    border-top: 1px solid $color-dfdeed;
-    border-right: 1px solid $color-dfdeed;
-    border-bottom: 1px solid $color-dfdeed;
-    background: $--b-main-white-color;
+    border-top: 1px solid #dfdeed;
+    border-right: 1px solid #dfdeed;
+    border-bottom: 1px solid #dfdeed;
+    background: #ffffff;
     padding: 6px 12px;
     border-radius: 0 6px 6px 0;
     font-family: 'Inter';
@@ -117,7 +112,7 @@ $color-dfdeed: #dfdeed;
     font-weight: 400;
     font-size: 13px;
     line-height: 16px;
-    color: $--b-main-gray-color;
+    color: #575775;
     min-width: 76px;
   }
 }
