+++ src/directives/index.js
@@ -1,0 +1,1 @@
+export { hintsDirective } from "./hintsDirective";
