+++ src/main.js
@@ -17,10 +17,12 @@ import App from './App.vue';
 import router from './router';
 import pinia from './plugins/pinia';
 import pkg from '../package';
-import './workers/map-worker/map.init';
 import { UA_LANGUAGE } from './i18n/ua';
-import { BlanballEventBus } from './workers/event-bus-worker';
-import { useHintsStore } from './stores/hints';
+
+import { hintsDirective } from './directives';
+
+import './guards/tech-works-page-guard'
+import './workers/map-worker/map.init';
 
 import './assets/styles/main.scss';
 import './assets/styles/normalize.scss';
@@ -52,28 +54,5 @@ app
   .use(Toast)
   .use(Maska)
   .use(createDeviceDetector())
-  .directive('hints', (el, hintsData) => {
-    const hintsStore = useHintsStore();
-
-    const hintsToShow = hintsData.value.filter((result) => {
-      return hintsStore.hintsData.results.some(
-        (value) => value.name === result.name
-      );
-    });
-
-    const calculatePosition = (hint) => {
-      return {
-        ...hint,
-        top: el.getBoundingClientRect().top + el.offsetHeight + hint.marginTop,
-        left: el.getBoundingClientRect().left - (hint.width / 2)
-      }
-    }
-
-    hintsToShow.map(calculatePosition);
-
-    if (hintsToShow.length) {
-      BlanballEventBus.emit('createHints', { hints: hintsToShow, element: el });
-    }
-  })
-
+  .directive('hints', hintsDirective)
   .mount('#app');
