+++ src/workers/event-bus-worker/event.bus.worker.js
@@ -4,26 +4,17 @@ export class EventBus {
   }
 
   on(eventName, fn) {
-    this.events[eventName] = this.events[eventName] || [];
-    this.events[eventName].push(fn);
+    this.events[eventName] = fn;
   }
 
-  off(eventName, fn) {
-    if (this.events[eventName]) {
-      for (let i = 0; i < this.events[eventName].length; i++) {
-        if (this.events[eventName][i] === fn) {
-          this.events[eventName].splice(i, 1);
-          break;
-        }
-      }
-    }
+  off(eventName) {
+    delete this.events[eventName];
   }
 
   emit(eventName, data) {
-    if (this.events[eventName]) {
-      this.events[eventName].forEach((fn) => {
-        fn(data);
-      });
+    const fn = this.events[eventName];
+    if (fn && typeof fn === 'function') {
+      fn(data);
     }
   }
 }
